mappers added

+++ event-notificator/src/main/java/dev/sorokin/eventnotificator/api/NotificationResponseDto.java
@@ -3,9 +3,9 @@ package dev.sorokin.eventnotificator.api;
 import dev.sorokin.eventnotificator.domain.NotificationPayload;
 
 public record NotificationResponseDto(
-        Integer notificationId,
+        Long notificationId,
         String type,
-        Integer eventId,
+        Long eventId,
         String createdAt,
         boolean isRead,
         String message,
