+++ event-notificator/src/main/java/dev/sorokin/eventnotificator/domain/NotificationResponse.java
@@ -1,9 +1,9 @@
 package dev.sorokin.eventnotificator.domain;
 
 public record NotificationResponse(
-        Integer notificationId,
+        Long notificationId,
         String type,
-        Integer eventId,
+        Long eventId,
         String createdAt,
         boolean isRead,
         String message,
